+++ htdocs/content/problem.php
@@ -1,23 +1,24 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <?php require_once "../modules/handler.php" ?>
+    <?php require_once "../modules/.php" ?>
     <title>Problem | Abscissa</title>
     <meta name="description" content="Abscissa dashboard page.">
     <meta name="keywords" content="abscissa dashboard">
     <!-- HEAD MODULES -->
-    <?php add("analytics") ?>
-    <?php add("settings") ?>
+    <?php add("h", "analytics") ?>
+    <?php add("h", "settings") ?>
     <!-- STYLE -->
     <link rel="stylesheet/less" type="text/css" href="../style/problem.less">
 </head>
 <body>
     <!-- WORKING MODULES -->
-    <?php add("katex") ?>
-    <?php add("less") ?>
+    <?php add("w", "katex") ?>
+    <?php add("w", "less") ?>
     <!-- UI MODULES -->
-    <?php add("navbar") ?>
-    <?php add("user") ?>
+    <?php add("i", "codemirror") ?>
+    <?php add("i", "navbar") ?>
+    <?php add("i", "user") ?>
     <!-- MAIN -->
     <div id="main">
         <div id="data">
@@ -25,10 +26,10 @@
             <div id="instructions"></div>
         </div>
         <div id="controls">
-            <textarea id="playground" class="input-textarea" placeholder="Your process might want to be here, in the playground."></textarea>
-            <p id="visor">$$ $$</p>
+            <textarea id="playground" class="input-textarea"></textarea>
+            <p id="visor"></p>
             <div>
-                <input type="text" class="input-text" id="answer" placeholder="Write your answer here.">
+                <input type="text" class="input-text" id="answer" placeholder="Write down your answer here." autocomplete="off" spellcheck="false">
                 <p id="result"></p>
                 <button class="input-button" id="validate">VALIDATE</button>
             </div>