+++ htdocs/modules/i-user.php
@@ -4,4 +4,4 @@
     <p id="userName"></p>
     <img src="../public/preferences.svg" alt="" class="user-icon" id="preferencesLogo">
 </div>
-<script src="../modules/js/user.js"></script>
+<script src="../modules/js/i-user.js"></script>