+++ htdocs/modules/js/w-katex.js
@@ -1,24 +1,26 @@
 // RENDER MATH FUNCTION
-function render(element) {
+function renderLaTeX(element) {
     renderMathInElement(element, {
         delimiters: [
             {left: "$$", right: "$$", display: true},
             {left: "$", right: "$", display: false}
-        ]
+        ],
+        strict: false,
+        throwOnError: false,
     });
 }
 
 // RENDERING
 document.addEventListener("DOMContentLoaded", function() {
     // INITIAL RENDERING
-    render(document.body);
+    renderLaTeX(document.body);
     // OBSERVER FOR DYNAMIC CHANGES RENDERING
-    const observer = new MutationObserver(function(mutationsList) {
-        for (const mutation of mutationsList) {
-            if (mutation.type === 'childList') {
-                mutation.addedNodes.forEach(node => {
+    const observer = new MutationObserver(function(changeList) {
+        for (const change of changeList) {
+            if (change.type === 'childList') {
+                change.addedNodes.forEach(node => {
                     if (node.nodeType === Node.ELEMENT_NODE) {
-                        render(node);
+                        renderLaTeX(node);
                     }
                 });
             }